+++ bcipy/signal/model/gaussian_mixture/__init__.py
@@ -1,0 +1,6 @@
+from .gaussian_mixture import GazeModelCombined, GazeModelIndividual
+
+__all__ = [
+    "GazeModelCombined",
+    "GazeModelIndividual",
+]