+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Our last release candidate before the official 2.0 release!
 - Multimodal Acquisition and Querying
     - Support for multiple devices in online querying #286
     - Support for trigger handling relative to a given device #293
+- Model
+    - Offline analysis of multimodal fusion. First version of gaze model is completed #294 Gaze reshaper is updated #294 Visualization for the gaze classification is added #294
 - Stimuli
     - Updates to ensure stimuli are presented at the same frequency #287
 - Dynamic Selection Window